Reworked to provide common NumericUpDown control which values are doubles.
