+++ samples/ControlCatalog/ControlCatalog.csproj
@@ -78,7 +78,7 @@
     <EmbeddedResource Include="Pages\MenuPage.xaml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
-    <EmbeddedResource Include="Pages\NumbersPage.xaml">
+    <EmbeddedResource Include="Pages\NumericUpDownPage.xaml">
       <SubType>Designer</SubType>
     </EmbeddedResource>
     <EmbeddedResource Include="Pages\ProgressBarPage.xaml">
@@ -173,8 +173,8 @@
     <Compile Include="Pages\ButtonSpinnerPage.xaml.cs">
       <DependentUpon>ButtonSpinnerPage.xaml</DependentUpon>
     </Compile>
-	<Compile Include="Pages\NumbersPage.xaml.cs">
-      <DependentUpon>NumbersPage.xaml</DependentUpon>
+	<Compile Include="Pages\NumericUpDownPage.xaml.cs">
+      <DependentUpon>NumericUpDownPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Pages\ScreenPage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
