--- src/Avalonia.Controls/NumericUpDown/AllowedSpecialValues.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace Avalonia.Controls
-{
-    [Flags]
-    public enum AllowedSpecialValues
-    {
-        None = 0,
-        NaN = 1,
-        PositiveInfinity = 2,
-        NegativeInfinity = 4,
-        AnyInfinity = PositiveInfinity | NegativeInfinity,
-        Any = NaN | AnyInfinity
-    }
-}
